# Possible serial driver optimisations