Hello,

I have made some modifications to Apple2's serial driver and would like to discuss them with you. There are four different optimisations in this pull request. 

The first one is passing --cpu 65c02 when building libsrc/apple2 for the apple2enh ; we then use this to save a single cycle on SER_GET, with sta(ptr1) instead of sta(ptr1,x). 

The second one is in SER_GET too, and consists of *not* trying to send data before receiving. Is this an ACIA 6551 requirement that there's nothing to send before receving? If so, as SER_PUT is not IRQ-driven, maybe we could get rid of the SendBuf and just make sure the byte is out before returning ?
This one shaves 8 cycles if there was nothing to send (and approx 10 bytes on disk/in RAM). I could also put the patch in the other instances of the same driver for the other platforms.

The third one is in SER_IRQ, and reworks the RecvFreeCnt decrement and comparison to do it in the X register. This shaves 4 cycles in the common and best case; in the worst case (where we assert flow control), it saves no cycle as we have to reload X to Index. This one is generalizable too to the other platforms.

Finally, I've defined the most frequently accessed variables, and the ones that are in the "firehose" codepath of receiving bytes, to point to 5 unused bytes in the zero page, thus saving about one cycle per access. This last one is not clean, but could be made clean with a new segment definition and everything, I suppose.

All of these are ideas. What do you think?

Thanks!

PS: This does not save enough cycles to handle 19200 bps : (the ROM IRQ handler does about 240 cycles per call, the ProDOS IRQ handler about 250, and callirq about 42). Plus 54 in SER_IRQ, it's about 600 cycles per IRQ, but getting the data out of ser_get() (about 43 cycles) and into a large buffer still makes it too slow.