@colinleroy 2023-08-23
I'll try! I remember I had a hard time at the beginning, but it was due to serial shenanigans, not the driver.