@oliverschmidt 2023-08-26
First of all thanks for your interest in (further) contributing to cc65!

1. The changes in the Makefile are supposed to be unnecessary as the `-t apple2enh` implies setting the 65C02 CPU.

2. I don't see us accessing "some" ZP variables directly from "some" source file. You write that they are unused. When next week someone does a similar change in some other source code, he might claim the same. I think that things don't work that way. From that perspective, using "some" ZP variables from within an interrupt handler is the worst case as the usage is unpredictable.