@colinleroy 2023-08-26
> First of all thanks for your interest in (further) contributing to cc65!

And thank you for the support in this journey!

>     1. The changes in the Makefile are supposed to be unnecessary as the `-t apple2enh` implies setting the 65C02 CPU.

Oh indeed, it does :)

>     2. I don't see us accessing "some" ZP variables directly from "some" source file. You write that they are unused. When next week someone does a similar change in some other source code, he might claim the same. I think that things don't work that way. From that perspective, using "some" ZP variables from within an interrupt handler is the worst case as the usage is unpredictable.

Yes, that's the fishy one. That's why I wrote that if it is deemed a good idea, it should be cleaned up with a new segment definition, so that they're not "some" variables anymore. But even then, we couldn't be sure whether someone, somewhere, already utilizes those 5 bytes in their own code. So, probably the best is for me to revert that part.

I hope the rest is at least somewhat good :)