@oliverschmidt 2023-08-26
As you are working so intensively with that code, I ask you this:

My understanding is that the 6551 doesn't only generate an interrupt when there's a new byte available but also when a hardware handshake line changes. But the serial driver code filters those interrupts and thinks that the interrupt wasn't caused by the 6551.

This leads as far as I can see to an "unclaimed interrupt". The more recent ProDOS versions allow for 255 unclaimed interrupts be halting with a fatal error. So my assumption is that after >255 hardware handshake line changes, the system dies.

If you are willing and able to spend the time and effort to test this scenario, it would great to learn if my concern is correct.