@colinleroy 2023-08-26
> My understanding is that the 6551 doesn't only generate an interrupt when there's a new byte available but also when a hardware handshake line changes. But the serial driver code filters those interrupts and thinks that the interrupt wasn't caused by the 6551.

I have noticed that we handle the interrupt only if we have a byte to fetch, yes (ACIA_STATUS & 0x08). And also that we're supposed to check the bit 7 of the status register - so I tried to do that at some point, with no success: bit 7 was always 0. According to http://csbruce.com/cbm/ftp/reference/swiftlink.txt, "bit #7 is not consistently set.  At speeds
under 9600 bps, bit #7 seems to work as designed." So we can't use this if we want to handle 9600bps (and frankly it would be so sad to stay at 4800 :sweat:).
 
> This leads as far as I can see to an "unclaimed interrupt". The more recent ProDOS versions allow for 255 unclaimed interrupts be halting with a fatal error. So my assumption is that after >255 hardware handshake line changes, the system dies.

I suppose this HW handshake change should come from the other side, then ? I'll try that.