@colinleroy 2023-08-27
> As you are working so intensively with that code, I ask you this:
> 
> My understanding is that the 6551 doesn't only generate an interrupt when there's a new byte available but also when a hardware handshake line changes. But the serial driver code filters those interrupts and thinks that the interrupt wasn't caused by the 6551.

So, as far as my tests went, no IRQs are fired when the line status changes. I've tested both when changing flow control from the Apple 2 itself (with the attached patch, and attached test file).

When changing it from "outside", this time with a patch storing 'Z' in the ring buffer instead of returning from SER_IRQ when ACIA_STATUS&0x08 == 0, and plugging/unplugging the serial adapter, nothing is inserted in RecvBuf and SER_GET still return SER_ERR_NO_DATA.

I had to make this second test this way because my emulation chain doesn't handle CRTSCTS, so I did it on bare metal. Fro the first test, I ran it under MAME and verified the absence of SER_IRQ calls in the trace
[flow.patch.gz](https://github.com/cc65/cc65/files/12448611/flow.patch.gz)
[test.c.gz](https://github.com/cc65/cc65/files/12448612/test.c.gz)
log.