@polluks2 2023-08-27
By the way do you know https://sites.google.com/site/drjohnbmatthews/apple2/ssc?