@oliverschmidt 2023-08-28
> SwiftLink-232 Application Notes

Yes, I know that document and especially the section on the interrupt bit.

> So, as far as my tests went, no IRQs are fired when the line status changes.

:-( That doesn't match any documentation - incl. the one on the SwiftLink-232...

> plugging/unplugging the serial adapter

That doesn't seem like a good way to test things to me. My understanding is that you use the SSC to connect your //c to a Raspberry Pi. My assumption is that you have the HSKi pin (https://old.pinouts.ru/SerialPorts/apple_iic_serial_pinout.shtml) connected to an output pin on the Pi. To which one?

If you look at pages 4 and 5 of http://www.applelogic.org/files/IICSCHEMATIC.pdf (and additionally take https://www.onsemi.com/pdf/datasheet/mc1488-d.pdf) into account you'll see that HSKi is connected to the 6551 DCD pin (not the DSR pin as one would assume!) so that's what is supposed to see changes.