@colinleroy 2023-08-28
> :-( That doesn't match any documentation - incl. the one on the SwiftLink-232...

Indeed. I'll try again at 1200bps to check if it's the same kind of issue as bit 7 of status register not raising to 1 when an IRQ fired.
 
> > plugging/unplugging the serial adapter
> 
> That doesn't seem like a good way to test things to me. My understanding is that you use the SSC to connect your //c to a Raspberry Pi. My assumption is that you have the HSKi pin (https://old.pinouts.ru/SerialPorts/apple_iic_serial_pinout.shtml) connected to an output pin on the Pi. To which one?
> 
> If you look at pages 4 and 5 of http://www.applelogic.org/files/IICSCHEMATIC.pdf (and additionally take https://www.onsemi.com/pdf/datasheet/mc1488-d.pdf) into account you'll see that HSKi is connected to the 6551 DCD pin (not the DSR pin as one would assume!) so that's what is supposed to see changes.

Sadly, I can't really tell you - I'm using a monster cable (that I know passes handshake status correctly, as I don't lose bytes on the Apple2 whereas I do in certain conditions on the emulator) which I'm not prepared to rip apart...
![image](https://github.com/cc65/cc65/assets/1016317/3aee2661-6170-4e8f-9655-1abc0131e71b)