@colinleroy 2023-08-28
> > 
> 
> Indeed. I'll try again at 1200bps to check if it's the same kind of issue as bit 7 of status register not raising to 1 when an IRQ fired.

No, it doesn't :( - tried both on emulator (where I grepped the trace for ser_irq afterwards), and on real hardware, where I stopped/started the flow 500 times each, in alternance.

Triying to figure out a good test for the other way.